+++ src/FillBorders.cpp
@@ -1,14 +1,15 @@
-#include <memory>
-#include <cmath>
 #include <algorithm>
+#include <array>
+#include <cmath>
+#include <memory>
 
 #include "avisynth.h"
 
 template <typename T>
 AVS_FORCEINLINE void memset16(T* ptr, T value, size_t num)
 {
-    while (num-- > 0)
-        *ptr++ = value;
+    while (--num > 0)
+        *++ptr = value;
 }
 
 template<typename T, typename T1>
@@ -18,7 +19,7 @@ AVS_FORCEINLINE auto lerp(const T1 fill, const T1 src, const int pos, const int 
         return std::clamp(((fill * 256 * pos / size) + (src * 256 * (size - pos) / size)) >> 8, 0, 255);
     else if constexpr (std::is_same_v<T, uint16_t>)
     {
-        const int64_t max_range = 1LL << bits;
+        const int64_t max_range{ 1LL << bits };
         return static_cast<int>(std::clamp(((fill * max_range * pos / size) + (src * max_range * (static_cast<int64_t>(size) - pos) / size)) >> bits, static_cast<int64_t>(0), max_range - 1));
     }
     else
@@ -27,12 +28,12 @@ AVS_FORCEINLINE auto lerp(const T1 fill, const T1 src, const int pos, const int 
 
 class FillBorders : public GenericVideoFilter
 {
-    int m_left;
-    int m_top;
-    int m_right;
-    int m_bottom;
+    std::array<int, 4> m_left;
+    std::array<int, 4> m_top;
+    std::array<int, 4> m_right;
+    std::array<int, 4> m_bottom;
     int m_mode;
-    int process[3];
+    std::array<int, 4> process;
     bool has_at_least_v8;
 
     template<typename T, typename T1>
@@ -41,7 +42,7 @@ class FillBorders : public GenericVideoFilter
     PVideoFrame(FillBorders::* processing)(PVideoFrame frame, IScriptEnvironment* env);
 
 public:
-    FillBorders(PClip _child, int left, int top, int right, int bottom, int mode, int y, int u, int v, IScriptEnvironment* env);
+    FillBorders(PClip _child, AVSValue left, AVSValue top, AVSValue right, AVSValue bottom, int mode, int y, int u, int v, int a, IScriptEnvironment* env);
     PVideoFrame __stdcall GetFrame(int n, IScriptEnvironment* env);
     int __stdcall SetCacheHints(int cachehints, int frame_range)
     {
@@ -53,60 +54,53 @@ template<typename T, typename T1>
 PVideoFrame FillBorders::fill(PVideoFrame frame, IScriptEnvironment* env)
 {
     env->MakeWritable(&frame);
-    const int size = sizeof(T);
-    const int planes_y[4] = { PLANAR_Y, PLANAR_U, PLANAR_V, PLANAR_A };
-    const int planes_r[4] = { PLANAR_G, PLANAR_B, PLANAR_R, PLANAR_A };
-    const int* planes = vi.IsRGB() ? planes_r : planes_y;
-    const int planecount = std::min(vi.NumComponents(), 3);
-    for (int i = 0; i < planecount; ++i)
+    constexpr std::array<int, 4> planes_y{ PLANAR_Y, PLANAR_U, PLANAR_V, PLANAR_A };
+    constexpr std::array<int, 4> planes_r{ PLANAR_G, PLANAR_B, PLANAR_R, PLANAR_A };
+    const int* planes{ vi.IsRGB() ? planes_r.data() : planes_y.data() };
+
+    for (int i{ 0 }; i < vi.NumComponents(); ++i)
     {
-        const int plane = planes[i];
-        const int height = frame->GetHeight(plane);
+        const int height{ frame->GetHeight(planes[i]) };
 
         if (process[i] == 3)
         {
-            const int stride = frame->GetPitch(plane) / size;
-            const int width = frame->GetRowSize(plane) / size;
-            T* dstp = reinterpret_cast<T*>(frame->GetWritePtr(plane));
-
-            const int left = m_left >> vi.GetPlaneWidthSubsampling(plane);
-            const int top = m_top >> vi.GetPlaneHeightSubsampling(plane);
-            const int right = m_right >> vi.GetPlaneWidthSubsampling(plane);
-            const int bottom = m_bottom >> vi.GetPlaneHeightSubsampling(plane);
+            const size_t stride{ frame->GetPitch(planes[i]) / sizeof(T) };
+            const size_t width{ frame->GetRowSize(planes[i]) / sizeof(T) };
+            T* dstp{ reinterpret_cast<T*>(frame->GetWritePtr(planes[i])) };
 
             switch (m_mode)
             {
                 case 0:
                     if constexpr ((std::is_same_v<T, uint8_t>))
                     {
-                        for (int y = top; y < height - bottom; ++y)
+                        for (int y{ m_top[i] }; y < height - m_bottom[i]; ++y)
                         {
-                            memset(dstp + stride * static_cast<int64_t>(y), (dstp + stride * static_cast<int64_t>(y))[left], left);
-                            memset(dstp + stride * static_cast<int64_t>(y) + width - right, (dstp + stride * static_cast<int64_t>(y) + width - right)[-1], right);
+                            memset(dstp + stride * static_cast<int64_t>(y), (dstp + stride * static_cast<int64_t>(y))[m_left[i]], m_left[i]);
+                            memset(dstp + stride * static_cast<int64_t>(y) + width - m_right[i], (dstp + stride * static_cast<int64_t>(y) + width - m_right[i])[-1], m_right[i]);
                         }
                     }
                     else
                     {
-                        for (int y = top; y < height - bottom; ++y)
+                        for (int y{ m_top[i] }; y < height - m_bottom[i]; ++y)
                         {
-                            memset16<T>(dstp + stride * static_cast<int64_t>(y), (dstp + stride * static_cast<int64_t>(y))[left], left);
-                            memset16<T>(dstp + stride * static_cast<int64_t>(y) + width - right, (dstp + stride * static_cast<int64_t>(y) + width - right)[-1], right);
+                            memset16<T>(dstp + stride * static_cast<int64_t>(y), (dstp + stride * static_cast<int64_t>(y))[m_left[i]], m_left[i]);
+                            memset16<T>(dstp + stride * static_cast<int64_t>(y) + width - m_right[i], (dstp + stride * static_cast<int64_t>(y) + width - m_right[i])[-1], m_right[i]);
                         }
                     }
 
-                    for (int y = top - 1; y >= 0; y--)
+                    for (int y{ m_top[i] - 1 }; y >= 0; --y)
                     {
                         // copy first pixel
                         // copy last eight pixels
                         dstp[stride * y] = dstp[stride * (y + 1)];
-                        memcpy(dstp + stride * static_cast<int64_t>(y) + width - 8, dstp + stride * (static_cast<int64_t>(y) + 1) + width - 8, 8LL * size);
+                        memcpy(dstp + stride * static_cast<int64_t>(y) + width - 8, dstp + stride * (static_cast<int64_t>(y) + 1) + width - 8, 8LL * sizeof(T));
 
                         // weighted average for the rest
-                        for (int x = 1; x < width - 8; ++x)
+                        for (int x{ 1 }; x < width - 8; ++x)
                         {
-                            T prev = dstp[stride * (y + 1) + x - 1];
-                            T cur = dstp[stride * (y + 1) + x];
-                            T next = dstp[stride * (y + 1) + x + 1];
+                            T prev{ dstp[stride * (y + 1) + x - 1] };
+                            T cur{ dstp[stride * (y + 1) + x] };
+                            T next{ dstp[stride * (y + 1) + x + 1] };
 
                             if constexpr (std::is_integral_v<T>)
                                 dstp[stride * y + x] = (3 * prev + 2 * cur + 3 * next + 4) / 8;
@@ -115,19 +109,19 @@ PVideoFrame FillBorders::fill(PVideoFrame frame, IScriptEnvironment* env)
                         }
                     }
 
-                    for (int y = height - bottom; y < height; ++y)
+                    for (int y{ height - m_bottom[i] }; y < height; ++y)
                     {
                         // copy first pixel
                         // copy last eight pixels
                         dstp[stride * y] = dstp[stride * (y - 1)];
-                        memcpy(dstp + stride * static_cast<int64_t>(y) + width - 8, dstp + stride * (static_cast<int64_t>(y) - 1) + width - 8, 8LL * size);
+                        memcpy(dstp + stride * static_cast<int64_t>(y) + width - 8, dstp + stride * (static_cast<int64_t>(y) - 1) + width - 8, 8LL * sizeof(T));
 
                         // weighted average for the rest
-                        for (int x = 1; x < width - 8; ++x)
+                        for (int x{ 1 }; x < width - 8; ++x)
                         {
-                            T prev = dstp[stride * (y - 1) + x - 1];
-                            T cur = dstp[stride * (y - 1) + x];
-                            T next = dstp[stride * (y - 1) + x + 1];
+                            T prev{ dstp[stride * (y - 1) + x - 1] };
+                            T cur{ dstp[stride * (y - 1) + x] };
+                            T next{ dstp[stride * (y - 1) + x + 1] };
 
                             if constexpr ((std::is_integral_v<T>))
                                 dstp[stride * y + x] = (3 * prev + 2 * cur + 3 * next + 4) / 8;
@@ -139,124 +133,124 @@ PVideoFrame FillBorders::fill(PVideoFrame frame, IScriptEnvironment* env)
                 case 1:
                     if constexpr ((std::is_same_v<T, uint8_t>))
                     {
-                        for (int y = top; y < height - bottom; ++y)
+                        for (int y{ m_top[i] }; y < height - m_bottom[i]; ++y)
                         {
-                            memset(dstp + stride * static_cast<int64_t>(y), (dstp + stride * static_cast<int64_t>(y))[left], left);
-                            memset(dstp + stride * static_cast<int64_t>(y) + width - right, (dstp + stride * static_cast<int64_t>(y) + width - right)[-1], right);
+                            memset(dstp + stride * static_cast<int64_t>(y), (dstp + stride * static_cast<int64_t>(y))[m_left[i]], m_left[i]);
+                            memset(dstp + stride * static_cast<int64_t>(y) + width - m_right[i], (dstp + stride * static_cast<int64_t>(y) + width - m_right[i])[-1], m_right[i]);
                         }
                     }
                     else
                     {
-                        for (int y = top; y < height - bottom; ++y)
+                        for (int y{ m_top[i] }; y < height - m_bottom[i]; ++y)
                         {
-                            memset16<T>(dstp + stride * static_cast<int64_t>(y), (dstp + stride * static_cast<int64_t>(y))[left], left);
-                            memset16<T>(dstp + stride * static_cast<int64_t>(y) + width - right, (dstp + stride * static_cast<int64_t>(y) + width - right)[-1], right);
+                            memset16<T>(dstp + stride * static_cast<int64_t>(y), (dstp + stride * static_cast<int64_t>(y))[m_left[i]], m_left[i]);
+                            memset16<T>(dstp + stride * static_cast<int64_t>(y) + width - m_right[i], (dstp + stride * static_cast<int64_t>(y) + width - m_right[i])[-1], m_right[i]);
                         }
                     }
 
-                    for (int y = 0; y < top; ++y)
-                        memcpy(dstp + static_cast<int64_t>(stride) * y, dstp + static_cast<int64_t>(stride) * top, static_cast<int64_t>(stride) * size);
+                    for (int y{ 0 }; y < m_top[i]; ++y)
+                        memcpy(dstp + static_cast<int64_t>(stride) * y, dstp + static_cast<int64_t>(stride) * m_top[i], static_cast<int64_t>(stride) * sizeof(T));
 
-                    for (int y = height - bottom; y < height; ++y)
-                        memcpy(dstp + static_cast<int64_t>(stride) * y, dstp + stride * (static_cast<int64_t>(height) - bottom - 1), static_cast<int64_t>(stride) * size);
+                    for (int y{ height - m_bottom[i] }; y < height; ++y)
+                        memcpy(dstp + static_cast<int64_t>(stride) * y, dstp + stride * (static_cast<int64_t>(height) - m_bottom[i] - 1), static_cast<int64_t>(stride) * sizeof(T));
                     break;
                 case 2:
-                    for (int y = top; y < height - bottom; ++y)
+                    for (int y{ m_top[i] }; y < height - m_bottom[i]; ++y)
                     {
-                        for (int x = 0; x < left; ++x)
-                            dstp[stride * y + x] = dstp[stride * y + left * 2 - 1 - x];
+                        for (int x{ 0 }; x < m_left[i]; ++x)
+                            dstp[stride * y + x] = dstp[stride * y + m_left[i] * 2 - 1 - x];
 
-                        for (int x = 0; x < right; ++x)
-                            dstp[stride * y + width - right + x] = dstp[stride * y + width - right - 1 - x];
+                        for (int x{ 0 }; x < m_right[i]; ++x)
+                            dstp[stride * y + width - m_right[i] + x] = dstp[stride * y + width - m_right[i] - 1 - x];
                     }
 
-                    for (int y = 0; y < top; ++y)
-                        memcpy(dstp + static_cast<int64_t>(stride) * y, dstp + stride * (top * 2LL - 1 - y), static_cast<int64_t>(stride) * size);
+                    for (int y{ 0 }; y < m_top[i]; ++y)
+                        memcpy(dstp + static_cast<int64_t>(stride) * y, dstp + stride * (m_top[i] * 2LL - 1 - y), static_cast<int64_t>(stride) * sizeof(T));
 
-                    for (int y = 0; y < bottom; ++y)
-                        memcpy(dstp + stride * (static_cast<int64_t>(height) - bottom + static_cast<int64_t>(y)), dstp + stride * (static_cast<int64_t>(height) - bottom - 1 - y), static_cast<int64_t>(stride) * size);
+                    for (int y{ 0 }; y < m_bottom[i]; ++y)
+                        memcpy(dstp + stride * (static_cast<int64_t>(height) - m_bottom[i] + static_cast<int64_t>(y)), dstp + stride * (static_cast<int64_t>(height) - m_bottom[i] - 1 - y), static_cast<int64_t>(stride) * sizeof(T));
                     break;
                 case 3:
-                    for (int y = top; y < height - bottom; ++y)
+                    for (int y{ m_top[i] }; y < height - m_bottom[i]; ++y)
                     {
-                        for (int x = 0; x < left; ++x)
-                            dstp[stride * y + x] = dstp[stride * y + left * 2 - x];
+                        for (int x{ 0 }; x < m_left[i]; ++x)
+                            dstp[stride * y + x] = dstp[stride * y + m_left[i] * 2 - x];
 
-                        for (int x = 0; x < right; ++x)
-                            dstp[stride * y + width - right + x] = dstp[stride * y + width - right - 2 - x];
+                        for (int x{ 0 }; x < m_right[i]; ++x)
+                            dstp[stride * y + width - m_right[i] + x] = dstp[stride * y + width - m_right[i] - 2 - x];
                     }
 
-                    for (int y = 0; y < top; ++y)
-                        memcpy(dstp + static_cast<int64_t>(stride) * y, dstp + stride * (top * 2LL - y), static_cast<int64_t>(stride) * size);
+                    for (int y{ 0 }; y < m_top[i]; ++y)
+                        memcpy(dstp + static_cast<int64_t>(stride) * y, dstp + stride * (m_top[i] * 2LL - y), static_cast<int64_t>(stride) * sizeof(T));
 
-                    for (int y = 0; y < bottom; ++y)
-                        memcpy(dstp + stride * (static_cast<int64_t>(height) - bottom + static_cast<int64_t>(y)), dstp + stride * (static_cast<int64_t>(height) - bottom - 2 - y), static_cast<int64_t>(stride) * size);
+                    for (int y{ 0 }; y < m_bottom[i]; ++y)
+                        memcpy(dstp + stride * (static_cast<int64_t>(height) - m_bottom[i] + static_cast<int64_t>(y)), dstp + stride * (static_cast<int64_t>(height) - m_bottom[i] - 2 - y), static_cast<int64_t>(stride) * sizeof(T));
                     break;
                 case 4:
-                    for (int y = top; y < height - bottom; ++y)
+                    for (int y{ m_top[i] }; y < height - m_bottom[i]; ++y)
                     {
-                        for (int x = 0; x < left; ++x)
-                            dstp[stride * y + x] = dstp[stride * y + width - right - left + x];
+                        for (int x{ 0 }; x < m_left[i]; ++x)
+                            dstp[stride * y + x] = dstp[stride * y + width - m_right[i] - m_left[i] + x];
 
-                        for (int x = 0; x < right; ++x)
-                            dstp[stride * y + width - right + x] = dstp[stride * y + left + x];
+                        for (int x{ 0 }; x < m_right[i]; ++x)
+                            dstp[stride * y + width - m_right[i] + x] = dstp[stride * y + m_left[i] + x];
                     }
 
-                    for (int y = 0; y < top; ++y)
-                        memcpy(dstp + static_cast<int64_t>(stride) * y, dstp + static_cast<int64_t>(stride) * (static_cast<int64_t>(height) - bottom - top + y), static_cast<int64_t>(stride) * size);
+                    for (int y{ 0 }; y < m_top[i]; ++y)
+                        memcpy(dstp + static_cast<int64_t>(stride) * y, dstp + static_cast<int64_t>(stride) * (static_cast<int64_t>(height) - m_bottom[i] - m_top[i] + y), static_cast<int64_t>(stride) * sizeof(T));
 
-                    for (int y = 0; y < bottom; ++y)
-                        memcpy(dstp + stride * (static_cast<int64_t>(height) - bottom + static_cast<int64_t>(y)), dstp + static_cast<int64_t>(stride) * (top + static_cast<int64_t>(y)), static_cast<int64_t>(stride) * size);
+                    for (int y{ 0 }; y < m_bottom[i]; ++y)
+                        memcpy(dstp + stride * (static_cast<int64_t>(height) - m_bottom[i] + static_cast<int64_t>(y)), dstp + static_cast<int64_t>(stride) * (m_top[i] + static_cast<int64_t>(y)), static_cast<int64_t>(stride) * sizeof(T));
                     break;
                 case 5:
                 {
-                    const int bits = vi.BitsPerComponent();
-                    const int pl = (vi.IsRGB()) ? 0 : i;
+                    const int bits{ vi.BitsPerComponent() };
+                    const int pl{ (vi.IsRGB()) ? 0 : i };
 
-                    for (int y = 0; y < top; ++y)
+                    for (int y{ 0 }; y < m_top[i]; ++y)
                     {
-                        for (int x = 0; x < width; ++x)
-                            dstp[stride * y + x] = lerp<T, T1>(dstp[x], dstp[stride * y + x], top - y, top, bits, pl);
+                        for (int x{ 0 }; x < width; ++x)
+                            dstp[stride * y + x] = lerp<T, T1>(dstp[x], dstp[stride * y + x], m_top[i] - y, m_top[i], bits, pl);
                     }
 
-                    const int start_bottom = height - bottom;
-                    for (int y = start_bottom; y < height; ++y)
+                    const int start_bottom{ height - m_bottom[i] };
+                    for (int y{ start_bottom }; y < height; ++y)
                     {
-                        for (int x = 0; x < width; ++x)
-                            dstp[stride * y + x] = lerp<T, T1>(dstp[x], dstp[stride * y + x], y - start_bottom, bottom, bits, pl);
+                        for (int x{ 0 }; x < width; ++x)
+                            dstp[stride * y + x] = lerp<T, T1>(dstp[x], dstp[stride * y + x], y - start_bottom, m_bottom[i], bits, pl);
                     }
 
-                    const int start_right = width - right;
-                    for (int y = 0; y < height; ++y)
+                    const size_t start_right{ width - m_right[i] };
+                    for (int y{ 0 }; y < height; ++y)
                     {
-                        for (int x = 0; x < left; ++x)
-                            dstp[stride * y + x] = lerp<T, T1>(dstp[x], dstp[stride * y + x], left - x, left, bits, pl);
+                        for (int x{ 0 }; x < m_left[i]; ++x)
+                            dstp[stride * y + x] = lerp<T, T1>(dstp[x], dstp[stride * y + x], m_left[i] - x, m_left[i], bits, pl);
 
-                        for (int x = 0; x < right; ++x)
-                            dstp[stride * y + start_right + x] = lerp<T, T1>(dstp[x], dstp[stride * y + start_right + x], x, right, bits, pl);
+                        for (int x{ 0 }; x < m_right[i]; ++x)
+                            dstp[stride * y + start_right + x] = lerp<T, T1>(dstp[x], dstp[stride * y + start_right + x], x, m_right[i], bits, pl);
                     }
                     break;
                 }
                 case 6:
-                    for (int x = left - 1; x >= 0; x--)
+                    for (int x{ m_left[i] - 1 }; x >= 0; --x)
                     {
-                        // copy pixels until top + 3/bottom + 3
+                        // copy pixels until m_top[i] + 3/m_bottom[i] + 3
                         // this way we avoid darkened corners when all sides need filling
-                        for (int y = 0; y < top + 3; y++)
+                        for (int y{ 0 }; y < m_top[i] + 3; ++y)
                             dstp[stride * y + x] = dstp[stride * y + x + 1];
-                        for (int y = bottom + 3; y > 0; y--)
+                        for (int y{ m_bottom[i] + 3 }; y > 0; --y)
                             dstp[stride * (height - y) + x] = dstp[stride * (height - y) + x + 1];
 
                         // weighted average for the rest
-                        for (int y = top + 3; y < height - (bottom + 3); y++)
+                        for (int y{ m_top[i] + 3 }; y < height - (m_bottom[i] + 3); ++y)
                         {
-                            T prev = dstp[stride * (y - 1) + x + 1];
-                            T cur = dstp[stride * (y)+x + 1];
-                            T next = dstp[stride * (y + 1) + x + 1];
+                            T prev{ dstp[stride * (y - 1) + x + 1] };
+                            T cur{ dstp[stride * (y)+x + 1] };
+                            T next{ dstp[stride * (y + 1) + x + 1] };
 
-                            T ref_prev = dstp[stride * (y - 1) + x + 2];
-                            T ref_cur = dstp[stride * (y)+x + 2];
-                            T ref_next = dstp[stride * (y + 1) + x + 2];
+                            T ref_prev{ dstp[stride * (y - 1) + x + 2] };
+                            T ref_cur{ dstp[stride * (y)+x + 2] };
+                            T ref_next{ dstp[stride * (y + 1) + x + 2] };
 
                             T fill_prev, fill_cur, fill_next;
 
@@ -273,13 +267,13 @@ PVideoFrame FillBorders::fill(PVideoFrame frame, IScriptEnvironment* env)
                                 fill_next = (1 * prev + 3 * cur + 5 * next) / 9.0f;
                             }
 
-                            T blur_prev = (2 * ref_prev + ref_cur + dstp[stride * (y - 2) + x + 2]) / 4;
-                            T blur_next = (2 * ref_next + ref_cur + dstp[stride * (y + 2) + x + 2]) / 4;
+                            T blur_prev{ static_cast<T>((2 * ref_prev + ref_cur + dstp[stride * (y - 2) + x + 2]) / 4) };
+                            T blur_next{ static_cast<T>((2 * ref_next + ref_cur + dstp[stride * (y + 2) + x + 2]) / 4) };
 
-                            T diff_next = abs(ref_next - fill_cur);
-                            T diff_prev = abs(ref_prev - fill_cur);
-                            T thr_next = abs(ref_next - blur_next);
-                            T thr_prev = abs(ref_prev - blur_prev);
+                            T diff_next{ static_cast<T>(abs(ref_next - fill_cur)) };
+                            T diff_prev{ static_cast<T>(abs(ref_prev - fill_cur)) };
+                            T thr_next{ static_cast<T>(abs(ref_next - blur_next)) };
+                            T thr_prev{ static_cast<T>(abs(ref_prev - blur_prev)) };
 
                             if (diff_next > thr_next)
                             {
@@ -295,25 +289,25 @@ PVideoFrame FillBorders::fill(PVideoFrame frame, IScriptEnvironment* env)
                         }
                     }
 
-                    for (int x = width - right; x < width; x++)
+                    for (size_t x{ width - m_right[i] }; x < width; ++x)
                     {
-                        // copy pixels until top + 3/bottom + 3
+                        // copy pixels until m_top[i] + 3/m_bottom[i] + 3
                         // this way we avoid darkened corners when all sides need filling
-                        for (int y = 0; y < top + 3; y++)
+                        for (int y{ 0 }; y < m_top[i] + 3; ++y)
                             dstp[stride * y + x] = dstp[stride * y + x - 1];
-                        for (int y = bottom + 3; y > 0; y--)
+                        for (int y{ m_bottom[i] + 3 }; y > 0; --y)
                             dstp[stride * (height - y) + x] = dstp[stride * (height - y) + x - 1];
 
                         // weighted average for the rest
-                        for (int y = top + 3; y < height - (bottom + 3); y++)
+                        for (int y{ m_top[i] + 3 }; y < height - (m_bottom[i] + 3); ++y)
                         {
-                            T prev = dstp[stride * (y - 1) + x - 1];
-                            T cur = dstp[stride * (y)+x - 1];
-                            T next = dstp[stride * (y + 1) + x - 1];
+                            T prev{ dstp[stride * (y - 1) + x - 1] };
+                            T cur{ dstp[stride * (y)+x - 1] };
+                            T next{ dstp[stride * (y + 1) + x - 1] };
 
-                            T ref_prev = dstp[stride * (y - 1) + x - 2];
-                            T ref_cur = dstp[stride * (y)+x - 2];
-                            T ref_next = dstp[stride * (y + 1) + x - 2];
+                            T ref_prev{ dstp[stride * (y - 1) + x - 2] };
+                            T ref_cur{ dstp[stride * (y)+x - 2] };
+                            T ref_next{ dstp[stride * (y + 1) + x - 2] };
 
                             T fill_prev, fill_cur, fill_next;
 
@@ -330,13 +324,13 @@ PVideoFrame FillBorders::fill(PVideoFrame frame, IScriptEnvironment* env)
                                 fill_next = (1 * prev + 3 * cur + 5 * next) / 9.0f;
                             }
 
-                            T blur_prev = (2 * ref_prev + ref_cur + dstp[stride * (y - 2) + x - 2]) / 4;
-                            T blur_next = (2 * ref_next + ref_cur + dstp[stride * (y + 2) + x - 2]) / 4;
+                            T blur_prev{ static_cast<T>((2 * ref_prev + ref_cur + dstp[stride * (y - 2) + x - 2]) / 4) };
+                            T blur_next{ static_cast<T>((2 * ref_next + ref_cur + dstp[stride * (y + 2) + x - 2]) / 4) };
 
-                            T diff_next = abs(ref_next - fill_cur);
-                            T diff_prev = abs(ref_prev - fill_cur);
-                            T thr_next = abs(ref_next - blur_next);
-                            T thr_prev = abs(ref_prev - blur_prev);
+                            T diff_next{ static_cast<T>(abs(ref_next - fill_cur)) };
+                            T diff_prev{ static_cast<T>(abs(ref_prev - fill_cur)) };
+                            T thr_next{ static_cast<T>(abs(ref_next - blur_next)) };
+                            T thr_prev{ static_cast<T>(abs(ref_prev - blur_prev)) };
 
                             if (diff_next > thr_next)
                             {
@@ -352,23 +346,23 @@ PVideoFrame FillBorders::fill(PVideoFrame frame, IScriptEnvironment* env)
                         }
                     }
 
-                    for (int y = top - 1; y >= 0; y--)
+                    for (int y{ m_top[i] - 1 }; y >= 0; --y)
                     {
                         // copy first pixel
                         // copy last eight pixels
                         dstp[stride * y] = dstp[stride * (y + 1)];
-                        memcpy(dstp + stride * static_cast<int64_t>(y) + width - 8, dstp + stride * (y + 1LL) + width - 8, 8 * size);
+                        memcpy(dstp + stride * static_cast<int64_t>(y) + width - 8, dstp + stride * (y + 1LL) + width - 8, 8 * sizeof(T));
 
                         // weighted average for the rest
-                        for (int x = 1; x < width - 8; x++)
+                        for (int x{ 1 }; x < width - 8; ++x)
                         {
-                            T prev = dstp[stride * (y + 1) + x - 1];
-                            T cur = dstp[stride * (y + 1) + x];
-                            T next = dstp[stride * (y + 1) + x + 1];
+                            T prev{ dstp[stride * (y + 1) + x - 1] };
+                            T cur{ dstp[stride * (y + 1) + x] };
+                            T next{ dstp[stride * (y + 1) + x + 1] };
 
-                            T ref_prev = dstp[stride * (y + 2) + x - 1];
-                            T ref_cur = dstp[stride * (y + 2) + x];
-                            T ref_next = dstp[stride * (y + 2) + x + 1];
+                            T ref_prev{ dstp[stride * (y + 2) + x - 1] };
+                            T ref_cur{ dstp[stride * (y + 2) + x] };
+                            T ref_next{ dstp[stride * (y + 2) + x + 1] };
 
                             T fill_prev, fill_cur, fill_next;
 
@@ -385,13 +379,13 @@ PVideoFrame FillBorders::fill(PVideoFrame frame, IScriptEnvironment* env)
                                 fill_next = (1 * prev + 3 * cur + 5 * next) / 9.0f;
                             }
 
-                            T blur_prev = (2 * ref_prev + ref_cur + dstp[stride * (y + 2) + x - 2]) / 4;
-                            T blur_next = (2 * ref_next + ref_cur + dstp[stride * (y + 2) + x + 2]) / 4;
+                            T blur_prev{ static_cast<T>((2 * ref_prev + ref_cur + dstp[stride * (y + 2) + x - 2]) / 4) };
+                            T blur_next{ static_cast<T>((2 * ref_next + ref_cur + dstp[stride * (y + 2) + x + 2]) / 4) };
 
-                            T diff_next = abs(ref_next - fill_cur);
-                            T diff_prev = abs(ref_prev - fill_cur);
-                            T thr_next = abs(ref_next - blur_next);
-                            T thr_prev = abs(ref_prev - blur_prev);
+                            T diff_next{ static_cast<T>(abs(ref_next - fill_cur)) };
+                            T diff_prev{ static_cast<T>(abs(ref_prev - fill_cur)) };
+                            T thr_next{ static_cast<T>(abs(ref_next - blur_next)) };
+                            T thr_prev{ static_cast<T>(abs(ref_prev - blur_prev)) };
 
                             if (diff_next > thr_next)
                             {
@@ -407,23 +401,23 @@ PVideoFrame FillBorders::fill(PVideoFrame frame, IScriptEnvironment* env)
                         }
                     }
 
-                    for (int y = height - bottom; y < height; y++)
+                    for (int y{ height - m_bottom[i] }; y < height; ++y)
                     {
                         // copy first pixel
                         // copy last eight pixels
                         dstp[stride * y] = dstp[stride * (y - 1)];
-                        memcpy(dstp + stride * static_cast<int64_t>(y) + width - 8, dstp + stride * (y - 1LL) + width - 8, 8 * size);
+                        memcpy(dstp + stride * static_cast<int64_t>(y) + width - 8, dstp + stride * (y - 1LL) + width - 8, 8 * sizeof(T));
 
                         // weighted average for the rest
-                        for (int x = 1; x < width - 8; x++)
+                        for (int x{ 1 }; x < width - 8; ++x)
                         {
-                            T prev = dstp[stride * (y - 1) + x - 1];
-                            T cur = dstp[stride * (y - 1) + x];
-                            T next = dstp[stride * (y - 1) + x + 1];
+                            T prev{ dstp[stride * (y - 1) + x - 1] };
+                            T cur{ dstp[stride * (y - 1) + x] };
+                            T next{ dstp[stride * (y - 1) + x + 1] };
 
-                            T ref_prev = dstp[stride * (y - 2) + x - 1];
-                            T ref_cur = dstp[stride * (y - 2) + x];
-                            T ref_next = dstp[stride * (y - 2) + x + 1];
+                            T ref_prev{ dstp[stride * (y - 2) + x - 1] };
+                            T ref_cur{ dstp[stride * (y - 2) + x] };
+                            T ref_next{ dstp[stride * (y - 2) + x + 1] };
 
                             T fill_prev, fill_cur, fill_next;
 
@@ -440,13 +434,13 @@ PVideoFrame FillBorders::fill(PVideoFrame frame, IScriptEnvironment* env)
                                 fill_next = (1 * prev + 3 * cur + 5 * next) / 9.0f;
                             }
 
-                            T blur_prev = (2 * ref_prev + ref_cur + dstp[stride * (y - 2) + x - 2]) / 4;
-                            T blur_next = (2 * ref_next + ref_cur + dstp[stride * (y - 2) + x + 2]) / 4;
+                            T blur_prev{ static_cast<T>((2 * ref_prev + ref_cur + dstp[stride * (y - 2) + x - 2]) / 4) };
+                            T blur_next{ static_cast<T>((2 * ref_next + ref_cur + dstp[stride * (y - 2) + x + 2]) / 4) };
 
-                            T diff_next = abs(ref_next - fill_cur);
-                            T diff_prev = abs(ref_prev - fill_cur);
-                            T thr_next = abs(ref_next - blur_next);
-                            T thr_prev = abs(ref_prev - blur_prev);
+                            T diff_next{ static_cast<T>(abs(ref_next - fill_cur)) };
+                            T diff_prev{ static_cast<T>(abs(ref_prev - fill_cur)) };
+                            T thr_next{ static_cast<T>(abs(ref_next - blur_next)) };
+                            T thr_prev{ static_cast<T>(abs(ref_prev - blur_prev)) };
 
                             if (diff_next > thr_next)
                             {
@@ -466,57 +460,142 @@ PVideoFrame FillBorders::fill(PVideoFrame frame, IScriptEnvironment* env)
         }
         else if (process[i] == 2)
         {
-            const int stride = frame->GetPitch(plane);
+            const int stride{ frame->GetPitch(planes[i]) };
 
-            env->BitBlt(frame->GetWritePtr(plane), stride, frame->GetReadPtr(plane), stride, frame->GetRowSize(), height);
+            env->BitBlt(frame->GetWritePtr(planes[i]), stride, frame->GetReadPtr(planes[i]), stride, frame->GetRowSize(planes[i]), height);
         }
     }
 
     return frame;
 }
 
-FillBorders::FillBorders(PClip _child, int left, int top, int right, int bottom, int mode, int y, int u, int v, IScriptEnvironment* env)
-    : GenericVideoFilter(_child), m_left(left), m_top(top), m_right(right), m_bottom(bottom), m_mode(mode)
+FillBorders::FillBorders(PClip _child, AVSValue left, AVSValue top, AVSValue right, AVSValue bottom, int mode, int y, int u, int v, int a, IScriptEnvironment* env)
+    : GenericVideoFilter(_child), m_mode(mode), process{ 1, 1, 1, 1 }
 {
     if (!vi.IsPlanar())
         env->ThrowError("FillBorders: only planar formats are supported.");
-    if (m_left < 0)
-        env->ThrowError("FillBorders: left must be equal to or greater than 0.");
-    if (m_top < 0)
-        env->ThrowError("FillBorders: top must be equal to or greater than 0.");
-    if (m_right < 0)
-        env->ThrowError("FillBorders: right must be equal to or greater than 0.");
-    if (m_bottom < 0)
-        env->ThrowError("FillBorders: bottom must be equal to or greater than 0.");
-    if (m_mode < 0 || m_mode > 6)
-        env->ThrowError("FillBorders: invalid mode.");
-    if (m_mode == 0 || m_mode == 1 || m_mode == 5 || m_mode == 6)
+
+    const int sw{ (vi.IsY() || vi.IsRGB()) ? 0 : vi.GetPlaneWidthSubsampling(PLANAR_U) };
+    const int sh{ (vi.IsY() || vi.IsRGB()) ? 0 : vi.GetPlaneHeightSubsampling(PLANAR_U) };
+
+    const int num_left{ (left.Defined() ? left.ArraySize() : 0) };
+    if (num_left > vi.NumComponents())
+        env->ThrowError("FillBorders: more left given than there are planes");
+
+    for (int i{ 0 }; i < num_left; ++i)
+        m_left[i] = left[i].AsInt();
+
+    switch (num_left)
     {
-        if (vi.width < m_left + m_right || vi.width <= m_left || vi.width <= m_right || vi.width < m_top + m_bottom || vi.height <= m_top || vi.height <= m_bottom)
-            env->ThrowError("FillBorders: the input clip is too small or the borders are too big.");
+        case 0: m_left.fill(0); break;
+        case 1:
+        {
+            m_left[1] = m_left[2] = m_left[0] >> sw;
+            m_left[3] = m_left[0];
+            break;
+        }
+        case 2:
+        {
+            m_left[2] = m_left[1];
+            m_left[3] = m_left[0];
+            break;
+        }
+        case 3: m_left[3] = m_left[0]; break;
     }
-    else if (m_mode == 2 || m_mode == 3 || m_mode == 4)
+
+    const int num_top{ (top.Defined() ? top.ArraySize() : 0) };
+    if (num_top > vi.NumComponents())
+        env->ThrowError("FillBorders: more top given than there are planes");
+
+    for (int i{ 0 }; i < num_top; ++i)
+        m_top[i] = top[i].AsInt();
+
+    switch (num_top)
     {
-        if (vi.width < 2 * m_left || vi.width < 2 * m_right || vi.height < 2 * m_top || vi.height < 2 * m_bottom)
-            env->ThrowError("FillBorders: the input clip is too small or the borders are too big.");
+        case 0: m_top.fill(0); break;
+        case 1:
+        {
+            m_top[1] = m_top[2] = m_top[0] >> sh;
+            m_top[3] = m_top[0];
+            break;
+        }
+        case 2:
+        {
+            m_top[2] = m_top[1];
+            m_top[3] = m_top[0];
+            break;
+        }
+        case 3: m_top[3] = m_top[0]; break;
     }
+
+    const int num_right{ (right.Defined() ? right.ArraySize() : 0) };
+    if (num_right > vi.NumComponents())
+        env->ThrowError("FillBorders: more right given than there are planes");
+
+    for (int i{ 0 }; i < num_right; ++i)
+        m_right[i] = right[i].AsInt();
+
+    switch (num_right)
+    {
+        case 0: m_right.fill(0); break;
+        case 1:
+        {
+            m_right[1] = m_right[2] = m_right[0] >> sw;
+            m_right[3] = m_right[0];
+            break;
+        }
+        case 2:
+        {
+            m_right[2] = m_right[1];
+            m_right[3] = m_right[0];
+            break;
+        }
+        case 3: m_right[3] = m_right[0]; break;
+    }
+
+    const int num_bottom{ (bottom.Defined() ? bottom.ArraySize() : 0) };
+    if (num_bottom > vi.NumComponents())
+        env->ThrowError("FillBorders: more bottom given than there are planes");
+
+    for (int i{ 0 }; i < num_bottom; ++i)
+        m_bottom[i] = bottom[i].AsInt();
+
+    switch (num_bottom)
+    {
+        case 0: m_bottom.fill(0); break;
+        case 1:
+        {
+            m_bottom[1] = m_bottom[2] = m_bottom[0] >> sh;
+            m_bottom[3] = m_bottom[0];
+            break;
+        }
+        case 2:
+        {
+            m_bottom[2] = m_bottom[1];
+            m_bottom[3] = m_bottom[0];
+            break;
+        }
+        case 3: m_bottom[3] = m_bottom[0]; break;
+    }
+
     if (y < 1 || y > 3)
         env->ThrowError("FillBorders: y must be between 1..3.");
     if (u < 1 || u > 3)
         env->ThrowError("FillBorders: u must be between 1..3.");
     if (v < 1 || v > 3)
         env->ThrowError("FillBorders: v must be between 1..3.");
+    if (a < 1 || a > 3)
+        env->ThrowError("FillBorders: a must be between 1..3.");
 
-    has_at_least_v8 = true;
-    try { env->CheckVersion(8); }
-    catch (const AvisynthError&)
-    {
-        has_at_least_v8 = false;
-    }
+    const int chr_w{ vi.width >> sw };
+    const int chr_h{ vi.height >> sh };
+    const std::array<int, 4> w{ vi.width, chr_w, chr_w, vi.width };
+    const std::array<int, 4> h{ vi.height, chr_h, chr_h, vi.height };
 
-    const int planes[3] = { y, u, v };
-    const int planecount = std::min(vi.NumComponents(), 3);
-    for (int i = 0; i < planecount; ++i)
+    has_at_least_v8 = env->FunctionExists("propShow");
+
+    const std::array<int, 4> planes{ y, u, v, a };
+    for (int i{ 0 }; i < vi.NumComponents(); ++i)
     {
         if (vi.IsRGB())
             process[i] = 3;
@@ -527,6 +606,28 @@ FillBorders::FillBorders(PClip _child, int left, int top, int right, int bottom,
                 case 3: process[i] = 3; break;
                 case 2: process[i] = 2; break;
                 default: process[i] = 1; break;
+            }
+        }
+
+        if (process[i] == 3)
+        {
+            if (m_left[i] < 0)
+                env->ThrowError("FillBorders: left must be equal to or greater than 0.");
+            if (m_top[i] < 0)
+                env->ThrowError("FillBorders: top must be equal to or greater than 0.");
+            if (m_right[i] < 0)
+                env->ThrowError("FillBorders: right must be equal to or greater than 0.");
+            if (m_bottom[i] < 0)
+                env->ThrowError("FillBorders: bottom must be equal to or greater than 0.");
+            if (m_mode == 0 || m_mode == 1 || m_mode == 5 || m_mode == 6)
+            {
+                if (w[i] < m_left[i] + m_right[i] || w[i] <= m_left[i] || w[i] <= m_right[i] || h[i] < m_top[i] + m_bottom[i] || h[i] <= m_top[i] || h[i] <= m_bottom[i])
+                    env->ThrowError("FillBorders: the input clip is too small or the borders are too big.");
+            }
+            else if (m_mode == 2 || m_mode == 3 || m_mode == 4)
+            {
+                if (w[i] < 2 * m_left[i] || w[i] < 2 * m_right[i] || h[i] < 2 * m_top[i] || h[i] < 2 * m_bottom[i])
+                    env->ThrowError("FillBorders: the input clip is too small or the borders are too big.");
             }
         }
     }
@@ -541,11 +642,11 @@ FillBorders::FillBorders(PClip _child, int left, int top, int right, int bottom,
 
 PVideoFrame __stdcall FillBorders::GetFrame(int n, IScriptEnvironment* env)
 {
-    PVideoFrame frame = child->GetFrame(n, env);
+    PVideoFrame frame{ child->GetFrame(n, env) };
 
     if (has_at_least_v8)
     {
-        const AVSMap* props = env->getFramePropsRO(frame);
+        const AVSMap* props{ env->getFramePropsRO(frame) };
         if (env->propNumElements(props, "_FieldBased") > 0 && env->propGetInt(props, "_FieldBased", 0, nullptr) > 0)
             env->ThrowError("FillBorders: frame must be not interlaced.");
     }
@@ -555,16 +656,19 @@ PVideoFrame __stdcall FillBorders::GetFrame(int n, IScriptEnvironment* env)
 
 AVSValue __cdecl Create_FillBorders(AVSValue args, void* user_data, IScriptEnvironment* env)
 {
+    enum { Clip, Left, Top, Right, Bottom, Mode, Y, U, V, A };
+
     return new FillBorders(
-        args[0].AsClip(),
-        args[1].AsInt(0),
-        args[2].AsInt(0),
-        args[3].AsInt(0),
-        args[4].AsInt(0),
-        args[5].AsInt(0),
-        args[6].AsInt(3),
-        args[7].AsInt(3),
-        args[8].AsInt(3),
+        args[Clip].AsClip(),
+        args[Left],
+        args[Top],
+        args[Right],
+        args[Bottom],
+        args[Mode].AsInt(0),
+        args[Y].AsInt(3),
+        args[U].AsInt(3),
+        args[V].AsInt(3),
+        args[A].AsInt(3),
         env);
 }
 
@@ -625,7 +729,7 @@ const char* __stdcall AvisynthPluginInit3(IScriptEnvironment * env, const AVS_Li
 {
     AVS_linkage = vectors;
 
-    env->AddFunction("FillBorders", "c[left]i[top]i[right]i[bottom]i[mode]i[y]i[u]i[v]i", Create_FillBorders, 0);
+    env->AddFunction("FillBorders", "c[left]i*[top]i*[right]i*[bottom]i*[mode]i[y]i[u]i[v]i[a]i", Create_FillBorders, 0);
     env->AddFunction("FillMargins", "c[left]i[top]i[right]i[bottom]i[y]i[u]i[v]i", Create_FillMargins, 0);
     return "FillBorders";
 }